repo-healer/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7
pydantic-settings>=2.2
pydriller>=2.5
radon>=6.0
flake8>=7.0
scikit-learn>=1.4
joblib>=1.3
numpy>=1.26
pandas>=2.2
openai>=1.30
tenacity>=8.3
PyGithub>=2.3
GitPython>=3.1
structlog>=24.1
orjson>=3.10
redis>=5.0
aiofiles>=23.2
slowapi>=0.1.9
